+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Properly extract and burn in embedded text subtitles using Jellyfin, Emby and Plex libraries
 - Fix bug where deleting a channel would not remove its data from XMLTV
 - Fix colorspace filter for some files with invalid color metadata
+- Fix playback of external subtitles on Windows
 
 ### Changed
 - Remove duplicate items from smart collections before scheduling